Modiffied script to add and update gnss metadata fields
